Conversational AI Agent/requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# OpenAI API
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Web framework for UI
panel>=1.3.0

# Parameter validation
param>=1.12.0
pydantic>=2.0.0

# HTTP requests
requests>=2.25.0

# Wikipedia search
wikipedia>=1.4.0

# Data handling (optional but useful)
numpy>=1.21.0
pandas>=1.3.0

# Additional utilities
typing-extensions>=4.0.0